refactor: move payload encoding stuff to alt file
